+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fix assignments generating code multiple times
 - Fix `reg` being a proper alias of `logic` (#189)
 - Fix genvar initialization in for-generate (#192)
+- Fix assignment operators in genvar steps
 
 ## 0.10.0 - 2020-06-15
 ### Added